# src/order-service/internal/docs/swagger.yaml
basePath: /v1
definitions:
  dto.ConsumerCancelOrderRequest:
    properties:
      cancellation_reason:
        example: CHANGED_MIND
        type: string
    type: object
  dto.CreateOrderRequest:
    properties:
      amount:
        type: integer
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: 'order not found: id=550e8400-e29b-41d4-a716-446655440000'
        type: string
    type: object
  dto.OrderCreatedResponse:
    properties:
      OrderID:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  dto.OrderResponse:
    properties:
      amount:
        example: 19999
        type: integer
      cancellationReason:
        example: INSUFFICIENT_FUNDS
        type: string
      createdAt:
        example: "2025-12-23T09:19:23.458426Z"
        type: string
      description:
        example: This is a description
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        allOf:
        - $ref: '#/definitions/value_objects.OrderStatus'
        example: PENDING
      updatedAt:
        example: "2025-12-23T09:19:23.458426Z"
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  dto.OrdersResponse:
    properties:
      limit:
        type: integer
      orders:
        items:
          $ref: '#/definitions/dto.OrderResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.StaffCancelOrderRequest:
    properties:
      cancellation_reason:
        example: OUT_OF_STOCK
        type: string
    type: object
  dto.UpdateOrderStatusRequest:
    properties:
      status:
        example: ASSEMBLING
        type: string
    type: object
  value_objects.OrderStatus:
    enum:
    - PENDING
    - PAID
    - CANCELLED
    - ASSEMBLING
    - ASSEMBLED
    - DELIVERING
    - ISSUED
    type: string
    x-enum-varnames:
    - OrderPending
    - OrderPaid
    - OrderCanceled
    - OrderAssembling
    - OrderAssembled
    - OrderDelivering
    - OrderIssued
info:
  contact: {}
  description: Order microservice
  title: Order Service API
  version: "1.0"
paths:
  /orders:
    get:
      consumes:
      - application/json
      description: Returns paginated list of orders for the authenticated user
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrdersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all orders for user
      tags:
      - orders_customers
    post:
      consumes:
      - application/json
      description: Creates a new order, a message is sent to payment service
      parameters:
      - description: User ID (UUID)
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Request to create an order
        in: body
        name: order_request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OrderCreatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create an order
      tags:
      - orders_customers
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Returns order by UUID
      parameters:
      - description: User ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get order by id
      tags:
      - orders_customers
  /orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: 'Customer cancels order, body is optional. Possible reasons: CHANGED_MIND'
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User ID (UUID)
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: cancel request
        in: body
        name: cancel_request
        schema:
          $ref: '#/definitions/dto.ConsumerCancelOrderRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Cancels order (by customer)
      tags:
      - orders_customers
  /staff/orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: |-
        Cancels order, used by store/staff, simulation of jwt role with api key that can be found in .env
        Possible reasons: OUT_OF_STOCK, DELIVERY_UNAVAILABLE
      parameters:
      - description: Staff API key
        in: header
        name: X-Staff-API-Key
        required: true
        type: string
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: cancel request
        in: body
        name: cancel_request
        required: true
        schema:
          $ref: '#/definitions/dto.StaffCancelOrderRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Cancels order (by store)
      tags:
      - orders_staff
  /staff/orders/{id}/status:
    patch:
      consumes:
      - application/json
      description: |-
        Updates order status, used by staff, simulation of jwt role with api key that can be found in .env
        Possible update pipeline: PAID -> ASSEMBLING -> ASSEMBLED -> DELIVERING -> ISSUED
      parameters:
      - description: Staff API key
        in: header
        name: X-Staff-API-Key
        required: true
        type: string
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Order status update request
        in: body
        name: status_request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update order status (by store/stuff)
      tags:
      - orders_staff
schemes:
- http
- https
swagger: "2.0"